# docs/swagger.yaml
basePath: /v1
definitions:
...
  github_com_pickle_pw_monolith_internal_models_requests.ProfileLinkReq:
    properties:
      id:
        type: string
      name:
        type: string
      position:
        type: integer
      url:
        type: string
    type: object
...
  github_com_pickle_pw_monolith_internal_models_requests.UpdateProfileReq:
    properties:
      description:
        type: string
      displayName:
        type: string
      links:
        items:
          $ref: '#/definitions/github_com_pickle_pw_monolith_internal_models_requests.ProfileLinkReq'
        type: array
      username:
        type: string
    type: object
...
  github_com_pickle_pw_monolith_internal_types.ImagePreviewRes:
    properties:
      previewId:
        type: string
      previewUrl:
        type: string
    type: object
  github_com_pickle_pw_monolith_internal_types.ImageRes:
    properties:
      url:
        type: string
    type: object
...
  github_com_pickle_pw_monolith_internal_types.SupabaseWebhookPayload:
    properties:
      old_record:
        additionalProperties: {}
        type: object
      record:
        additionalProperties: {}
        type: object
      schema:
        type: string
      table:
        type: string
      type:
        type: string
    type: object
  github_com_pickle_pw_monolith_internal_types.UsernameValidationRes:
    properties:
      message:
        type: string
      valid:
        type: boolean
    type: object
...
host: api.pickle.pw
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is a Pickle server.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Pickle API
  version: "1.0"
paths:
...
  /v1/supabase-webhooks/users:
    post:
      consumes:
      - application/json
      description: Create profile webhook
      parameters:
      - description: Webhook payload
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.SupabaseWebhookPayload'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Create profile webhook
      tags:
      - profiles
  /v1/users/{userId}:
    get:
      consumes:
      - application/json
      description: Get profile by ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get profile by ID
      tags:
      - profiles
  /v1/users/{userId}/avatar:
    get:
      consumes:
      - application/json
      description: Get profile avatar URL
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.ImageRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get profile avatar URL
      tags:
      - profiles
...
  /v1/users/me:
    get:
      consumes:
      - application/json
      description: Get my profile
      parameters:
      - description: Avatar size
        in: query
        name: avatarSize
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get my profile
      tags:
      - profiles
    patch:
      consumes:
      - application/json
      description: Update my profile
      parameters:
      - description: Update profile request
        in: body
        name: updateProfileReq
        required: true
        schema:
          $ref: '#/definitions/github_com_pickle_pw_monolith_internal_models_requests.UpdateProfileReq'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Update my profile
      tags:
      - profiles
  /v1/users/me/avatar:
    get:
      consumes:
      - application/json
      description: Get my profile avatar URL
      parameters:
      - description: Size
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.ImageRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get my profile avatar URL
      tags:
      - profiles
    post:
      consumes:
      - application/json
      description: Upload avatar
      parameters:
      - description: File
        in: formData
        name: file
        required: true
        type: file
      - description: URL
        in: formData
        name: url
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.ImagePreviewRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Upload avatar
      tags:
      - profiles
...
  /v1/users/validate-username:
    get:
      consumes:
      - application/json
      description: Validate username
      parameters:
      - description: Username
        in: query
        name: username
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_pickle_pw_monolith_internal_types.UsernameValidationRes'
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Validate username
      tags:
      - profiles
swagger: "2.0"